counter_app/local_cicd/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
